+++ SharpQuake/Networking/Client/client_main.cs
@@ -451,7 +451,7 @@ namespace SharpQuake
 					ent.angles.Y = bobjrotate;
 
 				if ( ( ent.effects & EntityEffects.EF_BRIGHTFIELD ) != 0 )
-					Host.RenderContext.Particles.EntityParticles( Host.Client.cl.time, ent.origin );
+					Host.RenderContext.World.Particles.EntityParticles( Host.Client.cl.time, ent.origin );
 
 				if ( ( ent.effects & EntityEffects.EF_MUZZLEFLASH ) != 0 )
 				{
@@ -482,25 +482,25 @@ namespace SharpQuake
 				}
 
 				if ( ent.model.Flags.HasFlag( EntityFlags.Gib ) )
-					Host.RenderContext.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 2 );
+					Host.RenderContext.World.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 2 );
 				else if ( ent.model.Flags.HasFlag( EntityFlags.ZomGib ) )
-					Host.RenderContext.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 4 );
+					Host.RenderContext.World.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 4 );
 				else if ( ent.model.Flags.HasFlag( EntityFlags.Tracer ) )
-					Host.RenderContext.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 3 );
+					Host.RenderContext.World.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 3 );
 				else if ( ent.model.Flags.HasFlag( EntityFlags.Tracer2 ) )
-					Host.RenderContext.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 5 );
+					Host.RenderContext.World.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 5 );
 				else if ( ent.model.Flags.HasFlag( EntityFlags.Rocket ) )
 				{
-					Host.RenderContext.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 0 );
+					Host.RenderContext.World.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 0 );
 					var dl = AllocDlight( i );
 					dl.origin = ent.origin;
 					dl.radius = 200;
 					dl.die = ( Single ) cl.time + 0.01f;
 				}
 				else if ( ent.model.Flags.HasFlag( EntityFlags.Grenade ) )
-					Host.RenderContext.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 1 );
+					Host.RenderContext.World.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 1 );
 				else if ( ent.model.Flags.HasFlag( EntityFlags.Tracer3 ) )
-					Host.RenderContext.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 6 );
+					Host.RenderContext.World.Particles.RocketTrail( Host.Client.cl.time, ref oldorg, ref ent.origin, 6 );
 
 				ent.forcelink = false;
